+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-06-24 19:05:10
-// Commit: test(storage): implement widget integration tests for storage
+// Updated at: 2026-06-25 18:40:11
+// Commit: chore(storage): update dependencies and cleanup unused imports in storage
 
 class storageComponentState {
-  final String commitRef = '2275';
-  final String lastMessage = 'test(storage): implement widget integration tests for storage';
-  final DateTime timestamp = DateTime.parse('2026-06-24T19:05:10+05:00');
+  final String commitRef = '2320';
+  final String lastMessage = 'chore(storage): update dependencies and cleanup unused imports in storage';
+  final DateTime timestamp = DateTime.parse('2026-06-25T18:40:11+05:00');
 }